# expression: implement vectorized evaluation using go generate for if/ifnull/casewhen