<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

use go generate to implement vectorized evaluation for if/ifnull/casewhen.

### What is changed and how it works?

```
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenIntSig-VecBuiltinFunc-12         	  231482	      4654 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenIntSig-NonVecBuiltinFunc-12      	   51706	     23181 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenIntSig-VecBuiltinFunc#01-12      	  215451	      5610 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenIntSig-NonVecBuiltinFunc#01-12   	   35743	     33401 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenIntSig-VecBuiltinFunc#02-12      	  194908	      6251 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenIntSig-NonVecBuiltinFunc#02-12   	   38318	     31012 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenIntSig-VecBuiltinFunc#03-12      	  178933	      6738 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenIntSig-NonVecBuiltinFunc#03-12   	   34862	     31454 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenRealSig-VecBuiltinFunc-12        	  246494	      4770 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenRealSig-NonVecBuiltinFunc-12     	   51843	     23119 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenRealSig-VecBuiltinFunc#01-12     	  208749	      5726 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenRealSig-NonVecBuiltinFunc#01-12  	   44770	     26950 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenRealSig-VecBuiltinFunc#02-12     	  186452	      6311 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenRealSig-NonVecBuiltinFunc#02-12  	   41029	     28989 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenRealSig-VecBuiltinFunc#03-12     	  172772	      6906 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenRealSig-NonVecBuiltinFunc#03-12  	   38419	     31053 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDecimalSig-VecBuiltinFunc-12     	  205488	      5913 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDecimalSig-NonVecBuiltinFunc-12  	   48788	     26468 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDecimalSig-VecBuiltinFunc#01-12  	  140762	      8427 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDecimalSig-NonVecBuiltinFunc#01-12         	   42426	     29572 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDecimalSig-VecBuiltinFunc#02-12            	  132496	      9515 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDecimalSig-NonVecBuiltinFunc#02-12         	   33649	     34114 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDecimalSig-VecBuiltinFunc#03-12            	   88670	     11509 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDecimalSig-NonVecBuiltinFunc#03-12         	   35038	     35622 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenStringSig-VecBuiltinFunc-12                	  132127	      8967 ns/op	      48 B/op	       3 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenStringSig-NonVecBuiltinFunc-12             	   50223	     24412 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenStringSig-VecBuiltinFunc#01-12             	   86025	     14636 ns/op	      48 B/op	       3 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenStringSig-NonVecBuiltinFunc#01-12          	   36178	     34150 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenStringSig-VecBuiltinFunc#02-12             	   92600	     13147 ns/op	      80 B/op	       3 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenStringSig-NonVecBuiltinFunc#02-12          	   36616	     36981 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenStringSig-VecBuiltinFunc#03-12             	   71150	     17257 ns/op	      80 B/op	       3 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenStringSig-NonVecBuiltinFunc#03-12          	   27950	     40920 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenTimeSig-VecBuiltinFunc-12                  	  210492	      5531 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenTimeSig-NonVecBuiltinFunc-12               	   38874	     32549 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenTimeSig-VecBuiltinFunc#01-12               	  168430	      7494 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenTimeSig-NonVecBuiltinFunc#01-12            	   29169	     41717 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenTimeSig-VecBuiltinFunc#02-12               	  155646	      8019 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenTimeSig-NonVecBuiltinFunc#02-12            	   29702	     42130 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenTimeSig-VecBuiltinFunc#03-12               	  134184	      9641 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenTimeSig-NonVecBuiltinFunc#03-12            	   26431	     46014 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDurationSig-VecBuiltinFunc-12              	  241564	      5137 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDurationSig-NonVecBuiltinFunc-12           	   49935	     27403 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDurationSig-VecBuiltinFunc#01-12           	  205159	      5799 ns/op	      80 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDurationSig-NonVecBuiltinFunc#01-12        	   40534	     30580 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDurationSig-VecBuiltinFunc#02-12           	  163465	      6595 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDurationSig-NonVecBuiltinFunc#02-12        	   41030	     29929 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDurationSig-VecBuiltinFunc#03-12           	  151656	      7419 ns/op	     128 B/op	       4 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenDurationSig-NonVecBuiltinFunc#03-12        	   35379	     30917 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenJSONSig-VecBuiltinFunc-12                  	  126318	      9357 ns/op	      48 B/op	       3 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenJSONSig-NonVecBuiltinFunc-12               	   50948	     23931 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenJSONSig-VecBuiltinFunc#01-12               	   86773	     13879 ns/op	      48 B/op	       3 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenJSONSig-NonVecBuiltinFunc#01-12            	   38020	     32431 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenJSONSig-VecBuiltinFunc#02-12               	   80497	     14881 ns/op	      80 B/op	       3 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenJSONSig-NonVecBuiltinFunc#02-12            	   34675	     34156 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenJSONSig-VecBuiltinFunc#03-12               	   71059	     16972 ns/op	      80 B/op	       3 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinCaseWhenJSONSig-NonVecBuiltinFunc#03-12            	   32198	     38938 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullIntSig-VecBuiltinFunc-12                     	  649114	      1875 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullIntSig-NonVecBuiltinFunc-12                  	   61212	     19419 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullRealSig-VecBuiltinFunc-12                    	  659928	      1816 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullRealSig-NonVecBuiltinFunc-12                 	   64942	     18715 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullDecimalSig-VecBuiltinFunc-12                 	  328513	      3679 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullDecimalSig-NonVecBuiltinFunc-12              	   52654	     23119 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullStringSig-VecBuiltinFunc-12                  	   94730	     12630 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullStringSig-NonVecBuiltinFunc-12               	   41422	     30014 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullTimeSig-VecBuiltinFunc-12                    	  413376	      3100 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullTimeSig-NonVecBuiltinFunc-12                 	   27920	     42237 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullDurationSig-VecBuiltinFunc-12                	  579303	      2086 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullDurationSig-NonVecBuiltinFunc-12             	   61572	     19236 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullJSONSig-VecBuiltinFunc-12                    	   96963	     12196 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfNullJSONSig-NonVecBuiltinFunc-12                 	   49309	     24018 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfIntSig-VecBuiltinFunc-12                         	  327759	      3565 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfIntSig-NonVecBuiltinFunc-12                      	   37045	     32828 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfRealSig-VecBuiltinFunc-12                        	  337926	      3634 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfRealSig-NonVecBuiltinFunc-12                     	   42175	     28197 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfDecimalSig-VecBuiltinFunc-12                     	  224820	      5328 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfDecimalSig-NonVecBuiltinFunc-12                  	   39336	     30842 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfStringSig-VecBuiltinFunc-12                      	  104245	     11578 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfStringSig-NonVecBuiltinFunc-12                   	   31914	     38015 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfTimeSig-VecBuiltinFunc-12                        	  285519	      4422 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfTimeSig-NonVecBuiltinFunc-12                     	   23288	     52422 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfDurationSig-VecBuiltinFunc-12                    	  353846	      3490 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfDurationSig-NonVecBuiltinFunc-12                 	   40723	     29652 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfJSONSig-VecBuiltinFunc-12                        	   99319	     12129 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinControlFuncGenerated/builtinIfJSONSig-NonVecBuiltinFunc-12                     	   32206	     36945 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	118.804s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test